data-engineering-project-2024_3/hvs/requirements.txt
kafka-python==2.1.0
cassandra-driver==3.25.0


